+++ src/main/kotlin/org/agrfesta/k/kards/texasholdem/tournaments/IncreasingGamePayments.kt
@@ -4,24 +4,24 @@ import org.agrfesta.k.kards.texasholdem.rules.gameplay.GamePayments
 
 // info: https://pokersoup.com/tool/blindStructureCalculator
 
-class IncreasingGamePayments(private val structure: List<LevelPayments>, private val gamesPerLevel: Int): GamePayments {
+class IncreasingGamePayments(private val structure: List<LevelPayments>, private val gamesPerLevel: UInt): GamePayments {
 
     init {
-        require(gamesPerLevel >= 1) { "Unable to create an IncreasingGamePayments, gamesPerLevel=$gamesPerLevel" }
+        require(gamesPerLevel >= 1u) { "Unable to create an IncreasingGamePayments, gamesPerLevel=$gamesPerLevel" }
         require(structure.isNotEmpty()) { "Unable to create an IncreasingGamePayments from an empty structure" }
     }
 
-    private var games = 0
+    private var games = 0u
 
-    override fun sb(): Int = structure[level()].sb
-    override fun bb(): Int = structure[level()].bb
-    override fun ante(): Int? = structure[level()].ante
+    override fun sb(): UInt = structure[level()].sb
+    override fun bb(): UInt = structure[level()].bb
+    override fun ante(): UInt? = structure[level()].ante
 
-    private fun level(): Int = (games / gamesPerLevel).coerceAtMost(structure.size-1)
+    private fun level(): Int = (games.toInt() / gamesPerLevel.toInt()).coerceAtMost(structure.size-1)
 
     fun nextGame() {
         games++
     }
 }
 
-data class LevelPayments(val sb: Int, val bb: Int, val ante: Int? = null)
+data class LevelPayments(val sb: UInt, val bb: UInt, val ante: UInt? = null)
